packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "epaper-dithering"
version = "1.0.0"
description = "Dithering algorithms for e-paper/e-ink displays"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "g4bri3lDev", email = "admin@g4bri3l.de"}
]

keywords = ["epaper", "eink", "dithering", "image-processing", "display"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0"
]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.10"
]

[project.urls]
Homepage = "https://opendisplay.org"
Repository = "https://github.com/OpenDisplay-org/epaper-dithering"
Documentation = "https://github.com/OpenDisplay-org/epaper-dithering#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/epaper_dithering"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]